# optimize embedding forward cuda kernel lookup_table_v2,test=develop