<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
Adjust Grid and Group size of LookupTableV2
 - Profiled in Docker with nsys
 - Tesla V100-SXM2,    CUDA 10.1,    NVIDIA-SMI 440.33.01,    Driver Version: 440.33.01

| op/kernel  | inputshape  |type| size |is_sparse|  before|after|
|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
|LookupTableV2|[16,16]| fp16|[2,768]  |false|6980ns | 3232ns |
|LookupTableV2|[16,35,1]|fp32|[10000,1500]|false|56110ns|14880ns|
|LookupTableV2|[16,16]|fp32|[2,768]|false| 7140ns |3696ns|

- Impacts of different Grid and Group size

|configuration|inputshape|type|size|is_sparse|timing|
|:---:|:---:|:---:|:---:|:---:|:---:|
|80, 128x8|[16,16]|fp16|[2,768]|false|4288ns|
|80, 1024x1|[16,16]|fp16|[2,768]|false|3888ns|
|40, 256x4|[16,16]|fp16|[2,768]|false|4160ns|